# src/main/java/com/workmotion/employeemanagement/dto/NewEmployeeRequest.java
package com.workmotion.employeemanagement.dto;

import lombok.Data;

import javax.validation.constraints.NotBlank;

@Data
public class NewEmployeeRequest {
	@NotBlank
	private String username;
	@NotBlank
	private String name;
}